Added Temp calculation file and improved optimization
For more advanced temp calc I added a separate file for those funcitons since it is very complicated. Not fully functional yet just working through the parts.

Added some global optimization methods to optimization file. Using this file as a testing ground for the calculations leading up to the advanced temp calc. That's what the GiveLayerAbs and Q things are for

Added VLT calculation to pcecalc which gives range of availbale vlt values