Closer to working temp calculation, Changed import format, Trying to speed up optimization, Changed # of evaluations for VLT in wpv file
Temp calculation is closer to working. The fsolve function is not working as intended but the plots of temperature seem closer to true than previously. It's mostly a units problem and making sure variables are applied correctly. Still in progress.

Optimization is extremely slow and/or fails with 9+ layers. Tightening the bounds on TiO2 helped it to at least successfully complete but still took a long time. The VLT constraint slows it down considerably. 2+ hours.

Added incident angle function to PCECalc.

Changed import style so only used functions are imported. Potentially can speed up calculations and therefore speed up the optimization process. This was applied to Calc, optimization, and PVColor files.