VLT constraint has stopped working, Still messing with temperature calculation
VLT constraint has stopped working in the optimization routine. Cannot figure out why. Going to try an older version and see if it works there. Added a new VLT calculation that was supposed to speed things up. It works well but constraint fails with all version of VLT calculation. Like it is totally ignored.

Still working on temperature calculation. Rearranged the equations when setting up the system of equations. Hasn't fixed the problem. Not sure how to deal with it. Could be equation or the coding.